Refactor PageList
